+++ continuation/run_continuation.py
@@ -2,7 +2,7 @@ from supporting_classes import Bus, Line
 import numpy as np
 from supporting_methods import print_title1, print_title2, print_title3
 from newton_raphson_method.support import run_newton_raphson
-from continuation.continuation_support import Continuation
+from continuation.support import Continuation
 """
 Continuation settings
 """
